requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
pypdf==3.15.0
google-generativeai==0.1.0
langchain-community==0.1.0
pinecone-client==2.3.0
